Move Utils to Gremlin.Net.Process
